refactor: Move event::realmexit::handle_realm_exit() to rec::exit
Event module shows all types of event. (rmi, rsi, Context, etc)
So it's good to place RecExitReason under the event module.
But handling functions for the RecExit should be placed
under the 'rec' module.
